flask-api/requirements.txt
flask
pyyaml
flask-cors
docker